fix(egress): break the egress-verify/egress-boot-marker import cycle
coverageForWindow moved into egress-verify.ts (where listEgress already
lives); egress-boot-marker.ts now only exports appendBootMarker and
BOOT_MARKER_METHOD, and egress-verify.ts imports BOOT_MARKER_METHOD
one-way. Fixes an audit:boundaries no-circular violation.
